# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Drishti_IC
  description: |
    Backend APIs for Drishti IC
    
  version: 2.0.0
  contact:
    name: Backend APIs
servers:
  - url: http://localhost:8000
    description: Local Development Server
...
paths:
...
  /api/v1/ic/{part_number}/details:
    get:
      tags: [IC Database]
      summary: Get IC Specification
      operationId: getICDetails
      description: Fetch IC details from the Golden Record database.
      parameters:
        - name: part_number
          in: path
          required: true
          schema:
            type: string
          example: "LM555"
      responses:
        '200':
          description: IC found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICSpecification'
        '404':
          description: IC not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "IC_NOT_FOUND"
                message: "Part number 'XYZ999' not found in database."
                suggestion: "This IC may be in the sync queue or fake registry."
  /api/v1/ic/{part_number}/datasheet:
    get:
      tags: [IC Database]
      summary: Download IC Datasheet PDF
      operationId: getICDatasheet
      description: Stream the locally stored PDF datasheet for an IC.
      parameters:
        - name: part_number
          in: path
          required: true
          schema:
            type: string
          example: "LM555"
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Datasheet not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "DATASHEET_NOT_FOUND"
                message: "Datasheet for 'LM555' not available locally."
  /api/v1/ic/search:
    get:
      tags: [IC Database]
      summary: Search ICs (text search with filters)
      operationId: searchICs
      description: |
        Text search for ICs with optional filters and sorting.
        
        Use this when a user types into the search bar. Supports the same
        pagination, filtering, and sorting options as `/api/v1/ic/list`.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
          example: "555"
        - name: manufacturer
          in: query
          schema:
            type: string
          description: Filter by manufacturer (substring match)
          example: "Texas"
        - name: package_type
          in: query
          schema:
            type: string
          description: Filter by package type (e.g., DIP, SOIC, QFN)
          example: "SOIC"
        - name: min_pins
          in: query
          schema:
            type: integer
            minimum: 1
          description: Minimum pin count filter
          example: 8
        - name: max_pins
          in: query
          schema:
            type: integer
            minimum: 1
          description: Maximum pin count filter
          example: 64
        - name: sort_by
          in: query
          schema:
            $ref: '#/components/schemas/ICSortBy'
          description: Field to sort by
          example: "part_number"
        - name: sort_dir
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
          description: Sort direction
          example: "asc"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICSearchResult'
  /api/v1/ic/list:
    get:
      tags: [IC Database]
      summary: List ICs (browse with filters)
      operationId: listICs
      description: |
        Browse the IC catalog with pagination, filtering, and sorting.
        Use this for table views when no free-text search is provided.
      parameters:
        - name: manufacturer
          in: query
          schema:
            type: string
          description: Filter by manufacturer (substring match)
          example: "Microchip"
        - name: package_type
          in: query
          schema:
            type: string
          description: Filter by package type (e.g., DIP, SOIC, QFN)
          example: "QFN"
        - name: min_pins
          in: query
          schema:
            type: integer
            minimum: 1
          description: Minimum pin count filter
          example: 16
        - name: max_pins
          in: query
          schema:
            type: integer
            minimum: 1
          description: Maximum pin count filter
          example: 128
        - name: sort_by
          in: query
          schema:
            $ref: '#/components/schemas/ICSortBy'
          description: Field to sort by
          example: "updated_at"
        - name: sort_dir
          in: query
          schema:
            $ref: '#/components/schemas/SortDirection'
          description: Sort direction
          example: "desc"
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Page size
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Zero-based offset
      responses:
        '200':
          description: List results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ICSearchResult'
              example:
                results:
                  - part_number: "LM555CN"
                    manufacturer: "TI"
                    manufacturer_name: "Texas Instruments"
                    pin_count: 8
                    package_type: "DIP"
                    description: "Precision Timer IC"
                    has_datasheet: true
                  - part_number: "STM32F407"
                    manufacturer: "STM"
                    manufacturer_name: "STMicroelectronics"
                    pin_count: 100
                    package_type: "LQFP"
                    description: "ARM Cortex-M4 MCU"
                    has_datasheet: true
                total_count: 248
                limit: 20
                offset: 0

  # ================================================================
  # DATASHEET MANAGEMENT ENDPOINTS
  # ================================================================
...
  /api/v1/sync/start:
    post:
      tags: [Sync Operations]
      summary: Start Weekly Sync
      operationId: startSync
      description: |
        Triggers background scraping job for all queued ICs.
        Requires internet connection.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncStartRequest'
      responses:
        '202':
          description: Sync job started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJobInfo'
        '409':
          description: Sync already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "SYNC_IN_PROGRESS"
                message: "A sync job is already running."
                current_job_id: "660e8400-e29b-41d4-a716-446655440099"
                progress_percentage: 45
...
  /api/v1/sync/cancel:
    post:
      tags: [Sync Operations]
      summary: Cancel Sync Job
      operationId: cancelSync
      description: Stop the currently running sync job.
      responses:
        '200':
          description: Sync cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          description: No sync job running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/sync/history:
    get:
      tags: [Sync Operations]
      summary: Get Sync History
      operationId: getSyncHistory
      description: List past sync job results.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Sync history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncHistoryResult'

  # ================================================================
  # SETTINGS ENDPOINTS
  # ================================================================
...
components:
  schemas:
...
    ICSpecification:
      type: object
      nullable: true
      description: |
        Golden Record IC specification.
        Note: Same part_number can have multiple entries from different manufacturers.
        Unique constraint is on (part_number, manufacturer) composite key.
      properties:
        part_number:
          type: string
          description: IC part number (e.g., "LM555", "STM32L031K6")
        manufacturer:
          type: string
          enum: [STM, TI]
          description: Manufacturer code (enum)
        manufacturer_name:
          type: string
          description: Full manufacturer name (e.g., "Texas Instruments")
        pin_count:
          type: integer
          description: Number of physical pins
        package_type:
          type: string
          description: Package format (DIP, QFN, BGA, SOIC, TSSOP, LQFP, etc.)
        description:
          type: string
          description: Brief description of IC function
        datasheet_url:
          type: string
          description: Original web URL where datasheet was downloaded
        datasheet_path:
          type: string
          description: Local file path to stored PDF
        has_datasheet:
          type: boolean
          description: Whether datasheet PDF is available locally
        voltage_min:
          type: number
          format: float
          description: Minimum operating voltage (V)
        voltage_max:
          type: number
          format: float
          description: Maximum operating voltage (V)
        operating_temp_min:
          type: number
          format: float
          description: Minimum operating temperature (°C)
        operating_temp_max:
          type: number
          format: float
          description: Maximum operating temperature (°C)
        electrical_specs:
          type: object
          additionalProperties: true
          description: Additional electrical specifications (flexible JSON)
        source:
          type: string
          enum: [MANUAL, SCRAPED_STM, SCRAPED_TI, SCRAPED_MOUSER, SCRAPED_DIGIKEY, SCRAPED_ALLDATASHEET]
          description: Data source identifier
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ICSortBy:
      type: string
      enum: [part_number, manufacturer, pin_count, updated_at]
      description: Allowed sort fields for IC list/search endpoints.
    SortDirection:
      type: string
      enum: [asc, desc]
      description: Sort direction for list/search endpoints.
    ICSearchResult:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              part_number:
                type: string
              manufacturer:
                type: string
              manufacturer_name:
                type: string
              pin_count:
                type: integer
              package_type:
                type: string
              description:
                type: string
              has_datasheet:
                type: boolean
        total_count:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # --------------------------------------------------
    # DATASHEET MANAGEMENT SCHEMAS
    # --------------------------------------------------
...
    SyncStartRequest:
      type: object
      properties:
        max_items:
          type: integer
          description: Maximum items to process (optional)
        retry_failed:
          type: boolean
          default: true
          description: Retry previously failed items
    SyncJobInfo:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PROCESSING]
        message:
          type: string
        queue_size:
          type: integer
        estimated_time_minutes:
          type: integer
...
    SyncHistoryResult:
      type: object
      properties:
        sync_jobs:
          type: array
          items:
            type: object
            properties:
              job_id:
                type: string
                format: uuid
              status:
                type: string
                enum: [COMPLETED, ERROR, CANCELLED]
              started_at:
                type: string
                format: date-time
              completed_at:
                type: string
                format: date-time
              total_items:
                type: integer
              success_count:
                type: integer
              failed_count:
                type: integer
              fake_count:
                type: integer
        total_count:
          type: integer

    # --------------------------------------------------
    # SETTINGS SCHEMAS
    # --------------------------------------------------
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        suggestion:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true